fix(rename): renamed reference-data to external-api
