+++ external-api/src/external-api/currencies.ts
