+++ external-api/src/external-api/industry-sectors.ts
