+++ portal-api/api-tests/v1/deals/deals-status-submitted-updates-unsubmitted-issued-facilities.api-test.js
@@ -7,7 +7,7 @@ const completedDeal = require('../../fixtures/deal-fully-completed-issued-and-un
 const { as } = require('../../api')(app);
 const createFacilities = require('../../createFacilities');
 const api = require('../../../src/v1/api');
-const externalApis = require('../../../src/reference-data/api');
+const externalApis = require('../../../src/external-api/api');
 
 describe('PUT /v1/deals/:id/status - to `Submitted` - issued/unconditional facility submission details', () => {
   let aBarclaysMaker;
