+++ portal-api/src/external-api/__mocks__/api.js
