+++ portal-api/src/external-api/__mocks__/currencies.js
