+++ portal-api/src/external-api/__mocks__/industry-sectors.js
