+++ portal-api/src/external-api/currencies.js
