+++ portal-api/src/external-api/industry-sectors.js
