+++ portal-api/src/external-api/send-email.js
